Moving QuestionAnswerOrchestrator into single-turn
